orderedPainting.sh: Add initial Slurm support
This commit attempts to add Slurm support by leaving the code
unchanged as much as possible and just adds in extra parts to
strings when Slurm support is requested.

There is one conversion from using backticks to eval to stop
the dropping of some quote marks that results in sbatch trying
to interpret the arguments of the command to be run and failing.
The output that was being captured from the backticks wasn't
being used either as it was commented out, so I removed that.